+++ ShellResult.php
@@ -8,7 +8,7 @@
 namespace yii2tech\selfupdate;
 
 use yii\base\ErrorHandler;
-use yii\base\Object;
+use yii\base\BaseObject;
 
 /**
  * ShellResult represents shell command execution result.
@@ -18,7 +18,7 @@ use yii\base\Object;
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 1.0
  */
-class ShellResult extends Object
+class ShellResult extends BaseObject
 {
     /**
      * @var string command being executed.