+++ VersionControlSystem.php
@@ -7,7 +7,7 @@
 
 namespace yii2tech\selfupdate;
 
-use yii\base\Object;
+use yii\base\BaseObject;
 
 /**
  * VersionControlSystem is a base class for version control systems.
@@ -15,6 +15,6 @@ use yii\base\Object;
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 1.0
  */
-abstract class VersionControlSystem extends Object implements VersionControlSystemInterface
+abstract class VersionControlSystem extends BaseObject implements VersionControlSystemInterface
 {
 }